feat: 更新next-remoter组件，移除不必要的代码并添加提示功能
